module structure cleanup.
