+++ src/static_str.rs
@@ -1,0 +1,31 @@
+use std::{
+  borrow::Cow,
+  collections::HashSet,
+  sync::{OnceLock, PoisonError, RwLock},
+};
+
+pub type StaticStr = &'static str;
+pub type CowStr = Cow<'static, str>;
+
+/// Convert any str into a static str, using a global cache.
+#[inline]
+pub fn static_str(s: &str) -> StaticStr {
+  static STR_CACHE: OnceLock<RwLock<HashSet<StaticStr>>> = OnceLock::new();
+  let rw_lock = STR_CACHE.get_or_init(|| RwLock::new(HashSet::new()));
+  let read = rw_lock.read().unwrap_or_else(PoisonError::into_inner);
+  if let Some(out) = read.get(s) {
+    out
+  } else {
+    drop(read);
+    let mut write = rw_lock.write().unwrap_or_else(PoisonError::into_inner);
+    // It's *possible* that the string was inserted after we dropped the reader
+    // before we acquired the writer, so we check a second time.
+    if let Some(out) = write.get(s) {
+      out
+    } else {
+      let leaked: StaticStr = Box::leak(s.to_string().into_boxed_str());
+      write.insert(leaked);
+      leaked
+    }
+  }
+}
