+++ tests/parsing_tests.rs
@@ -3,7 +3,7 @@ use yagbas::{
   const_expr::ConstExpr,
   token::{tokenize_module, Token},
   token_tree::{grow_token_trees, TokenTree},
-  *,
+  util_junk::*,
 };
 use Token::*;
 use TokenTree::*;
